refactor: extract SubmissionGrader and move GraderService to dedicated package
- Create autograder.services.grader package.
- Relocate GraderService to services/grader/grader_service.py.
- Extract stateful tree traversal logic into SubmissionGrader within services/grader/criteria_grader.py.
- Update GraderService to act as a stateless coordinator delegating to SubmissionGrader.
- Refactor unit tests to use SubmissionGrader and update internal imports.
- Update project documentation to reflect structural changes.
