+++ docs/pipeline/04.5-ai-batch.md
@@ -36,7 +36,7 @@ If no `AiTestFunction` instances are found in the criteria tree, the step exits 
 
 ## How It Integrates with Grade
 
-The `GradeStep` reads the `AI_BATCH` step result from the pipeline and passes it as `pre_computed_results` to `GraderService.grade_from_tree()`. The grader threads this dict through the tree traversal into every `process_test()` call, which forwards it as a kwarg to `test_function.execute()`.
+The `GradeStep` reads the `AI_BATCH` step result from the pipeline and passes it as `pre_computed_results` to `GraderService.grade_from_tree()`. The `SubmissionGrader` threads this dict through the tree traversal into every `process_test()` call, which forwards it as a kwarg to `test_function.execute()`.
 
 `AiTestFunction.execute()` sees the `pre_computed_results` kwarg, looks up its own `name` in the dict, and returns the pre-computed `TestResult` directly — no further API call, no mutation.
